+++ Packages/com.noovertime.virtual-texture/CHANGELOG.md
@@ -5,4 +5,6 @@
 ###### Fixed
 
 - 修复GetVirtualPageID中解码packedImageInfo有误
+- 修复VirtualImageAtlas中计算父节点index有误
+- 修复部分shader编译警告